+++ app/src/main/java/io/familymoments/app/feature/postdetail/model/request/PostLovesRequest.kt
@@ -1,5 +1,5 @@
 package io.familymoments.app.feature.postdetail.model.request
 
 data class PostLovesRequest(
-    val postId:Int
+    val postId:Long
 )
